Dashboard, admin check moved
