+++ resources/js/store/store.ts
@@ -3,8 +3,6 @@ import {defineStore} from 'pinia';
 import {NewBugReport} from 'resources/types/bug';
 import {Error} from 'resources/types/error';
 import {NewFeedback} from 'resources/types/feedback';
-import {useRewardStore} from './rewardStore';
-import {useTaskStore} from './taskStore';
 
 export const useMainStore = defineStore('main', {
     state: () => {
@@ -22,14 +20,7 @@ export const useMainStore = defineStore('main', {
         },
         clearErrors() {
             this.errors = {};
-        },
-        async getDashboard() {
-            const {data} = await axios.get('/dashboard');
-            const taskStore = useTaskStore();
-            taskStore.taskLists = data.taskLists;
-            const rewardStore = useRewardStore();
-            rewardStore.rewardObj = data.rewardObj;
-        },
+        }, 
         async sendFeedback(feedback: NewFeedback) {
             await axios.get('/sanctum/csrf-cookie');
             await axios.post('/feedback', feedback);
